+++ internal/app/helpers/funcmap.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/rvflash/elapsed"
 	"github.com/ugent-library/biblio-backoffice/internal/models"
+	"github.com/ugent-library/friendly"
 )
 
 func FuncMap() template.FuncMap {
@@ -14,6 +15,7 @@ func FuncMap() template.FuncMap {
 		"timeElapsed": elapsed.LocalTime,
 		"formatRange": FormatRange,
 		"formatBool":  FormatBool,
+		"formatBytes": friendly.Bytes,
 	}
 }