+++ internal/app/routes/register.go
@@ -138,6 +138,7 @@ func Register(services *backends.Services, baseURL *url.URL, router *mux.Router,
 		BaseHandler: baseHandler,
 		Repository:  services.Repository,
 		FileStore:   services.FileStore,
+		MaxFileSize: viper.GetInt("max-file-size"),
 	}
 	publicationCreatingHandler := &publicationcreating.Handler{
 		BaseHandler:              baseHandler,
@@ -158,6 +159,7 @@ func Register(services *backends.Services, baseURL *url.URL, router *mux.Router,
 		PersonService:             services.PersonService,
 		DatasetSearchService:      services.DatasetSearchService,
 		FileStore:                 services.FileStore,
+		MaxFileSize:               viper.GetInt("max-file-size"),
 	}
 	publicationBatchHandler := &publicationbatch.Handler{
 		BaseHandler:    baseHandler,
@@ -1012,6 +1014,10 @@ func Register(services *backends.Services, baseURL *url.URL, router *mux.Router,
 		publicationEditingHandler.Wrap(publicationEditingHandler.EditFile)).
 		Methods("GET").
 		Name("publication_edit_file")
+	r.HandleFunc("/publication/{id}/refresh-files",
+		publicationEditingHandler.Wrap(publicationEditingHandler.RefreshFiles)).
+		Methods("GET").
+		Name("publication_refresh_files")
 	r.HandleFunc("/publication/{id}/files/{file_id}/refresh-form",
 		publicationEditingHandler.Wrap(publicationEditingHandler.RefreshEditFileForm)).
 		Methods("GET").