+++ internal/commands/publication.go
@@ -1,16 +1,18 @@
 package commands
 
 import (
+	"context"
 	"encoding/json"
 	"errors"
 	"io"
 	"log"
 	"os"
 	"strings"
-	"sync"
+	"time"
 
 	"github.com/oklog/ulid/v2"
 	"github.com/spf13/cobra"
+	"github.com/ugent-library/biblio-backoffice/internal/backends"
 	"github.com/ugent-library/biblio-backoffice/internal/models"
 	"github.com/ugent-library/biblio-backoffice/internal/snapstore"
 )
@@ -25,8 +27,9 @@ func init() {
 	publicationCmd.AddCommand(publicationAddCmd)
 	publicationCmd.AddCommand(publicationImportCmd)
 	publicationCmd.AddCommand(oldPublicationImportCmd)
-	publicationCmd.AddCommand(cleanupCmd)
-	publicationCmd.AddCommand(transferCmd)
+	publicationCmd.AddCommand(publicationCleanupCmd)
+	publicationCmd.AddCommand(publicationTransferCmd)
+	publicationCmd.AddCommand(publicationReindexCmd)
 	rootCmd.AddCommand(publicationCmd)
 }
 
@@ -88,19 +91,9 @@ var publicationAddCmd = &cobra.Command{
 	Use:   "add",
 	Short: "Add publications",
 	Run: func(cmd *cobra.Command, args []string) {
+		ctx := context.Background()
+
 		e := Services()
-
-		var indexWG sync.WaitGroup
-
-		// indexing channel
-		indexC := make(chan *models.Publication)
-
-		// start bulk indexer
-		indexWG.Add(1)
-		go func() {
-			defer indexWG.Done()
-			e.PublicationSearchService.IndexMultiple(indexC)
-		}()
 
 		fmt, _ := cmd.Flags().GetString("format")
 		decFactory, ok := e.PublicationDecoders[fmt]
@@ -109,15 +102,28 @@ var publicationAddCmd = &cobra.Command{
 		}
 		dec := decFactory(os.Stdin)
 
+		bi, err := e.PublicationSearchService.NewBulkIndexer(backends.BulkIndexerConfig{
+			OnError: func(err error) {
+				log.Printf("Indexing failed : %s", err)
+			},
+			OnIndexError: func(id string, err error) {
+				log.Printf("Indexing failed for publication [id: %s] : %s", id, err)
+			},
+		})
+		if err != nil {
+			log.Fatal(err)
+		}
+		defer bi.Close(ctx)
+
 		lineNo := 0
 		for {
 			lineNo += 1
-			p := models.Publication{
+			p := &models.Publication{
 				ID:             ulid.Make().String(),
 				Status:         "private",
 				Classification: "U",
 			}
-			if err := dec.Decode(&p); errors.Is(err, io.EOF) {
+			if err := dec.Decode(p); errors.Is(err, io.EOF) {
 				break
 			} else if err != nil {
 				log.Fatalf("Unable to decode publication at line %d : %v", lineNo, err)
@@ -126,17 +132,14 @@ var publicationAddCmd = &cobra.Command{
 				log.Printf("Validation failed for publication [id: %s] at line %d : %v", p.ID, lineNo, err)
 				continue
 			}
-			if err := e.Repository.SavePublication(&p, nil); err != nil {
+			if err := e.Repository.SavePublication(p, nil); err != nil {
 				log.Fatalf("Unable to store publication from line %d : %v", lineNo, err)
 			}
 
-			indexC <- &p
+			if err := bi.Index(ctx, p); err != nil {
+				log.Printf("Indexing failed for publication [id: %s] at line %d : %s", p.ID, lineNo, err)
+			}
 		}
-
-		// close indexing channel when all recs are stored
-		close(indexC)
-		// wait for indexing to finish
-		indexWG.Wait()
 	},
 }
 
@@ -144,19 +147,9 @@ var publicationImportCmd = &cobra.Command{
 	Use:   "import",
 	Short: "Import publications",
 	Run: func(cmd *cobra.Command, args []string) {
+		ctx := context.Background()
+
 		e := Services()
-
-		var indexWG sync.WaitGroup
-
-		// indexing channel
-		indexC := make(chan *models.Publication)
-
-		// start bulk indexer
-		indexWG.Add(1)
-		go func() {
-			defer indexWG.Done()
-			e.PublicationSearchService.IndexMultiple(indexC)
-		}()
 
 		fmt, _ := cmd.Flags().GetString("format")
 		decFactory, ok := e.PublicationDecoders[fmt]
@@ -165,11 +158,24 @@ var publicationImportCmd = &cobra.Command{
 		}
 		dec := decFactory(os.Stdin)
 
+		bi, err := e.PublicationSearchService.NewBulkIndexer(backends.BulkIndexerConfig{
+			OnError: func(err error) {
+				log.Printf("Indexing failed : %s", err)
+			},
+			OnIndexError: func(id string, err error) {
+				log.Printf("Indexing failed for publication [id: %s] : %s", id, err)
+			},
+		})
+		if err != nil {
+			log.Fatal(err)
+		}
+		defer bi.Close(ctx)
+
 		lineNo := 0
 		for {
 			lineNo += 1
-			p := models.Publication{}
-			if err := dec.Decode(&p); errors.Is(err, io.EOF) {
+			p := &models.Publication{}
+			if err := dec.Decode(p); errors.Is(err, io.EOF) {
 				break
 			} else if err != nil {
 				log.Fatalf("Unable to decode publication at line %d : %v", lineNo, err)
@@ -184,7 +190,7 @@ var publicationImportCmd = &cobra.Command{
 				)
 				continue
 			}
-			if err := e.Repository.ImportCurrentPublication(&p); err != nil {
+			if err := e.Repository.ImportCurrentPublication(p); err != nil {
 				log.Printf(
 					"Unable to store publication[snapshot_id: %s, id: %s] from line %d : %v",
 					p.SnapshotID,
@@ -200,36 +206,35 @@ var publicationImportCmd = &cobra.Command{
 				p.ID,
 			)
 
-			indexC <- &p
+			if err := bi.Index(ctx, p); err != nil {
+				log.Printf("Indexing failed for publication [id: %s] : %s", p.ID, err)
+			}
 		}
-
-		// close indexing channel when all recs are stored
-		close(indexC)
-		// wait for indexing to finish
-		indexWG.Wait()
 	},
 }
 
-var cleanupCmd = &cobra.Command{
+var publicationCleanupCmd = &cobra.Command{
 	Use:   "cleanup",
 	Short: "Make publications consistent, clean up data anomalies",
 	Run: func(cmd *cobra.Command, args []string) {
-		s := newRepository()
+		ctx := context.Background()
+
 		e := Services()
 
-		var indexWG sync.WaitGroup
+		bi, err := e.PublicationSearchService.NewBulkIndexer(backends.BulkIndexerConfig{
+			OnError: func(err error) {
+				log.Printf("Indexing failed : %s", err)
+			},
+			OnIndexError: func(id string, err error) {
+				log.Printf("Indexing failed for publication [id: %s] : %s", id, err)
+			},
+		})
+		if err != nil {
+			log.Fatal(err)
+		}
+		defer bi.Close(ctx)
 
-		// indexing channel
-		indexC := make(chan *models.Publication)
-
-		// start bulk indexer
-		indexWG.Add(1)
-		go func() {
-			defer indexWG.Done()
-			e.PublicationSearchService.IndexMultiple(indexC)
-		}()
-
-		s.EachPublication(func(p *models.Publication) bool {
+		e.Repository.EachPublication(func(p *models.Publication) bool {
 			// Guard
 			fixed := false
 
@@ -273,7 +278,7 @@ var cleanupCmd = &cobra.Command{
 					return false
 				}
 
-				err := s.UpdatePublication(p.SnapshotID, p, nil)
+				err := e.Repository.UpdatePublication(p.SnapshotID, p, nil)
 
 				var conflict *snapstore.Conflict
 				if errors.As(err, &conflict) {
@@ -292,20 +297,17 @@ var cleanupCmd = &cobra.Command{
 					p.ID,
 				)
 
-				indexC <- p
+				if err := bi.Index(ctx, p); err != nil {
+					log.Printf("Indexing failed for publication [id: %s] : %s", p.ID, err)
+				}
 			}
 
 			return true
 		})
-
-		// close indexing channel when all recs are stored
-		close(indexC)
-		// wait for indexing to finish
-		indexWG.Wait()
 	},
 }
 
-var transferCmd = &cobra.Command{
+var publicationTransferCmd = &cobra.Command{
 	Use:   "transfer UID UID [PUBID]",
 	Short: "Transfer publications between people",
 	Args:  cobra.RangeArgs(2, 3),
@@ -472,5 +474,94 @@ var oldPublicationImportCmd = &cobra.Command{
 				p.ID,
 			)
 		}
+	},
+}
+
+var publicationReindexCmd = &cobra.Command{
+	Use:   "reindex",
+	Short: "Reindex into a new search index",
+	Run: func(cmd *cobra.Command, args []string) {
+		ctx := context.Background()
+
+		services := Services()
+
+		startTime := time.Now()
+
+		indexed := 0
+
+		log.Println("Indexing to new index...")
+
+		switcher, err := services.PublicationSearchService.NewIndexSwitcher(backends.BulkIndexerConfig{
+			OnError: func(err error) {
+				log.Printf("Indexing failed : %s", err)
+			},
+			OnIndexError: func(id string, err error) {
+				log.Printf("Indexing failed for publication [id: %s] : %s", id, err)
+			},
+		})
+		if err != nil {
+			log.Fatal(err)
+		}
+		services.Repository.EachPublication(func(p *models.Publication) bool {
+			if err := switcher.Index(ctx, p); err != nil {
+				log.Printf("Indexing failed for publication [id: %s] : %s", p.ID, err)
+			}
+			indexed++
+			return true
+		})
+
+		log.Printf("Indexed %d publications...", indexed)
+
+		log.Println("Switching to new index...")
+
+		if err := switcher.Switch(ctx); err != nil {
+			log.Fatal(err)
+		}
+
+		endTime := time.Now()
+
+		log.Println("Indexing changes since start of reindex...")
+
+		for {
+			indexed = 0
+
+			bi, err := services.PublicationSearchService.NewBulkIndexer(backends.BulkIndexerConfig{
+				OnError: func(err error) {
+					log.Printf("Indexing failed : %s", err)
+				},
+				OnIndexError: func(id string, err error) {
+					log.Printf("Indexing failed for publication [id: %s] : %s", id, err)
+				},
+			})
+			if err != nil {
+				log.Fatal(err)
+			}
+
+			err = services.Repository.PublicationsBetween(startTime, endTime, func(p *models.Publication) bool {
+				if err := bi.Index(ctx, p); err != nil {
+					log.Printf("Indexing failed for publication [id: %s] : %s", p.ID, err)
+				}
+				indexed++
+				return true
+			})
+			if err != nil {
+				log.Fatal(err)
+			}
+
+			if err = bi.Close(ctx); err != nil {
+				log.Fatal(err)
+			}
+
+			if indexed == 0 {
+				break
+			}
+
+			log.Printf("Indexed %d publications...", indexed)
+
+			startTime = endTime
+			endTime = time.Now()
+		}
+
+		log.Println("Done.")
 	},
 }