+++ CHANGELOG.md
@@ -9,7 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- File upload progress
+- Show error dialog if upload is too large
+- Make maximum file size configurable
+- Reindex command
+- Better WOS Type facet
+- Search on organization ID
+
 ### Fixed
+
+- Fix conflict error after file upload cancel
 
 ## [v1.0.21] - 2023-01-25